Add user authentication

+++ CleanArchitecture.Domain/Commands/Users/DeleteUser/DeleteUserCommandHandler.cs
@@ -1,5 +1,6 @@
 using System.Threading;
 using System.Threading.Tasks;
+using CleanArchitecture.Domain.Enums;
 using CleanArchitecture.Domain.Errors;
 using CleanArchitecture.Domain.Events.User;
 using CleanArchitecture.Domain.Interfaces;
@@ -12,15 +13,18 @@ namespace CleanArchitecture.Domain.Commands.Users.DeleteUser;
 public sealed class DeleteUserCommandHandler : CommandHandlerBase,
     IRequestHandler<DeleteUserCommand>
 {
+    private readonly IUser _user;
     private readonly IUserRepository _userRepository;
-    
+
     public DeleteUserCommandHandler(
         IMediatorHandler bus,
         IUnitOfWork unitOfWork,
         INotificationHandler<DomainNotification> notifications,
-        IUserRepository userRepository) : base(bus, unitOfWork, notifications)
+        IUserRepository userRepository,
+        IUser user) : base(bus, unitOfWork, notifications)
     {
         _userRepository = userRepository;
+        _user = user;
     }
 
     public async Task Handle(DeleteUserCommand request, CancellationToken cancellationToken)
@@ -31,7 +35,7 @@ public sealed class DeleteUserCommandHandler : CommandHandlerBase,
         }
 
         var user = await _userRepository.GetByIdAsync(request.UserId);
-        
+
         if (user == null)
         {
             await NotifyAsync(
@@ -43,11 +47,22 @@ public sealed class DeleteUserCommandHandler : CommandHandlerBase,
             return;
         }
 
+        if (_user.GetUserId() != request.UserId && _user.GetUserRole() != UserRole.Admin)
+        {
+            await NotifyAsync(
+                new DomainNotification(
+                    request.MessageType,
+                    $"No permission to delete user {request.UserId}",
+                    ErrorCodes.InsufficientPermissions));
+
+            return;
+        }
+
         _userRepository.Remove(user);
 
         if (await CommitAsync())
         {
-            await _bus.RaiseEventAsync(new UserDeletedEvent(request.UserId));
+            await Bus.RaiseEventAsync(new UserDeletedEvent(request.UserId));
         }
     }
 }
